// src/app/services/models/register-model.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Competence } from '../models/competence';
export interface RegisterModel {
  competences?: Array<Competence> | null;
  cvFile?: File | null;
  email?: string | null;
  firstname?: string | null;
  lastname?: string | null;
  password?: string | null;
  telephone?: string | null;
}